feat: add global Supabase module and service with env validation
