+++ apps/backend/src/app.module.ts
@@ -4,6 +4,7 @@ import { appConfig, validate } from './config/app.config'
 import { ExampleModule } from './modules/example/example.module'
 import { PrismaModule } from './common/prisma/prisma.module'
 import { AuthModule } from './modules/auth/auth.module'
+import { SupabaseModule } from './modules/supabase/supabase.module'
 
 @Module({
   imports: [
@@ -16,6 +17,7 @@ import { AuthModule } from './modules/auth/auth.module'
     PrismaModule,
     ExampleModule,
     AuthModule,
+    SupabaseModule,
   ],
 })
 export class AppModule {}
