+++ apps/backend/src/modules/supabase/supabase.module.ts
@@ -1,0 +1,10 @@
+import { Global, Module } from '@nestjs/common'
+import { SupabaseService } from './supabase.service'
+
+@Global()
+@Module({
+  imports: [],
+  providers: [SupabaseService],
+  exports: [SupabaseService],
+})
+export class SupabaseModule {}
